Fixes #29717 - Fix React tests
Recently nock was introduced to testing. Importing `nock` from a place we also use `axios-mock-adapter` was causing issues as these mock network requests differently. Nock mocks from the nodejs level and axios-mock-adapter mocks the axios library. So initializing axios-mock-adapter and nock caused axios-mock-adapter to try to interecept the request before nock could.

We didn't see this locally because our dev env is on nodejs 10 and jenkins is on 12. I'm not sure why some runs passed and others have not.

Its likely a snapshot has to be re-recorded as well for the SubscriptionActions failure.
