+++ webpack/scenes/Subscriptions/__tests__/SubscriptionsActions.test.js
@@ -48,7 +48,7 @@ const store = mockStore(Immutable({
   },
 }));
 
-afterEach(() => {
+beforeEach(() => {
   store.clearActions();
   mockReset();
 });
